Move expand_gene_set to wdae utils module
